refactor: lift state up to Game component for centralized game logic management
