+++ app/game/page.tsx
@@ -3,26 +3,62 @@
 import GameHeader from "../ui/game/game-header";
 import GameBoard from "../ui/game/game-board";
 import StartGameTimer from "../ui/game/start-game-timer";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 
 export default function Game() {
     const [isTimerRunning, setIsTimerRunning] = useState(false);
+
+    const [time, setTime] = useState(0);
+
+    useEffect(() => {
+        let timer: NodeJS.Timeout | null = null;
+
+        if (isTimerRunning) {
+            timer = setInterval(() => {
+                setTime((prevTime) => prevTime + 1);
+            }, 1000);
+        } else if (!isTimerRunning && timer) {
+            clearInterval(timer);
+        }
+
+        return () => {
+            if (timer) clearInterval(timer);
+        };
+    }, [isTimerRunning]);
+
+    const [score, setScore] = useState(2500);
+    const [eggsFound, setEggsFound] = useState(0);
+    const [gameOver, setGameOver] = useState(false);
+
+    const handleCellClick = (item: string) => {
+        if (gameOver) return; // Prevent further clicks if the game is over
+
+        if (item === "🥚") {
+            setScore((prevScore) => prevScore + 100);
+            setEggsFound((prevEggsFound) => prevEggsFound + 1);
+        } else {
+            setScore((prevScore) => prevScore - 100);
+        }
+    };
+
+    useEffect(() => {
+        if (eggsFound === 5) {
+            setGameOver(true); // End the game when all 5 eggs are found
+            setIsTimerRunning(false); // Stop the timer
+        }
+    }, [eggsFound]);
 
     const handleStart = () => {
         setIsTimerRunning(true);
     };
 
-    const stopTimer = () => {
-        setIsTimerRunning(false);
-    };
-
     return (
         <>
-            <GameHeader timerIsRunning={isTimerRunning} />
+            <GameHeader timer={time} />
             <div className="flex flex-col items-center justify-center min-h-screen bg-gray-100">
                 <h1 className="text-4xl font-bold mb-4">Egg Hunt Game</h1>
                 <p className="text-lg mb-8">Uncover hidden eggs in the grid!</p>
-                <GameBoard onGameCellClick={handleStart} onGameOver={stopTimer} />
+                <GameBoard onGameCellClick={handleStart} handleCellClick={handleCellClick} gameOver={gameOver} score={score} />
                 <StartGameTimer onStart={handleStart} />
             </div>
         </>
