+++ app/ui/game/game-board.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import GameCell from "./game-cell";
 
 const items = [
@@ -15,31 +15,10 @@ function shuffleArray(array: string[]) {
     return array.sort(() => Math.random() - 0.5);
 }
 
-export default function GameBoard({ onGameCellClick, onGameOver }: { onGameCellClick: () => void, onGameOver: () => void }) {
-    const [score, setScore] = useState(2500);
+export default function GameBoard({ onGameCellClick, handleCellClick, gameOver, score }: { onGameCellClick: () => void, handleCellClick: (item: string) => void, gameOver: boolean, score: number }) {
     const [gameBoard] = useState(() =>
         shuffleArray(items).map((item) => ({ item, isRevealed: false }))
     );
-    const [eggsFound, setEggsFound] = useState(0);
-    const [gameOver, setGameOver] = useState(false);
-
-    const handleCellClick = (item: string) => {
-        if (gameOver) return; // Prevent further clicks if the game is over
-
-        if (item === "🥚") {
-            setScore((prevScore) => prevScore + 100);
-            setEggsFound((prevEggsFound) => prevEggsFound + 1);
-        } else {
-            setScore((prevScore) => prevScore - 100);
-        }
-    };
-
-    useEffect(() => {
-        if (eggsFound === 5) {
-            setGameOver(true); // End the game when all 5 eggs are found
-            onGameOver(); // Call the end game function
-        }
-    }, [eggsFound, onGameOver]);
 
     return (
         <div>
@@ -48,7 +27,7 @@ export default function GameBoard({ onGameCellClick, onGameOver }: { onGameCellC
                     <GameCell
                         key={index}
                         item={cell.item}
-                        onClick={() => handleCellClick(cell.item)}
+                        onClick={handleCellClick.bind(null, cell.item)}
                         startTimer={onGameCellClick}
                         disabled={gameOver || cell.isRevealed} // Disable cell if game is over or already revealed
                     />
